requirements.txt

requests
python-dotenv
PyJWT
jsonschema
pytest
